refactor(configaudit): split rc/pip audits into internal/detector/configaudit
Moves the 15 npmrc + pipconfig files out of internal/detector/ and into
a dedicated sub-package internal/detector/configaudit/. Pure rename + a
package-decl change in each file (`package detector` → `package configaudit`)
plus updated imports in the three callsites:

- internal/scan/scanner.go
- internal/telemetry/telemetry.go
- cmd/stepsecurity-dev-machine-guard/main.go

`detector.NewNPMRCDetector(...)` becomes `configaudit.NewNPMRCDetector(...)`
and likewise for `NewPipConfigDetector`. No public-API changes beyond the
package qualifier.

Why: internal/detector/ had grown to 55 files and the rc/config-file
audits are a self-contained subset (15 files, no cross-references with
sibling detector files). Splitting them off drops detector/ to 40 files
and clarifies the architecture: detector/ owns inventory-style detectors
(IDEs, AI tools, package managers, etc.); configaudit/ owns config-file
audits with their own discovery + parsing + finding semantics. Future
sibling audits — .yarnrc.yml, .gemrc, .cargo/config.toml — fit naturally
under configaudit/.

Verified: full test suite green (configaudit tests run as a separate
package now), cross-compile clean for darwin/linux/windows.
