+++ internal/detector/configaudit/npmrc_parse.go
@@ -1,4 +1,4 @@
-package detector
+package configaudit
 
 import (
 	"bufio"
