+++ internal/detector/configaudit/npmrc_stat_windows.go
@@ -1,6 +1,6 @@
 //go:build windows
 
-package detector
+package configaudit
 
 // statOwner is a no-op on Windows: getting a meaningful owner string from a
 // SID is non-trivial and not actionable for the audit's first cut. The
