+++ internal/detector/configaudit/npmrc_test.go
@@ -1,4 +1,4 @@
-package detector
+package configaudit
 
 import (
 	"context"
