+++ internal/detector/configaudit/pipconfig_findings.go
@@ -1,4 +1,4 @@
-package detector
+package configaudit
 
 import (
 	"fmt"
