+++ internal/detector/configaudit/pipconfig_helpers.go
@@ -1,4 +1,4 @@
-package detector
+package configaudit
 
 import (
 	"crypto/sha256"
