+++ internal/scan/scanner.go
@@ -8,6 +8,7 @@ import (
 	"github.com/step-security/dev-machine-guard/internal/buildinfo"
 	"github.com/step-security/dev-machine-guard/internal/cli"
 	"github.com/step-security/dev-machine-guard/internal/detector"
+	"github.com/step-security/dev-machine-guard/internal/detector/configaudit"
 	"github.com/step-security/dev-machine-guard/internal/device"
 	"github.com/step-security/dev-machine-guard/internal/executor"
 	"github.com/step-security/dev-machine-guard/internal/model"
@@ -212,14 +213,14 @@ func Run(exec executor.Executor, log *progress.Logger, cfg *cli.Config) error {
 	log.StepStart("Auditing npm configuration")
 	start = time.Now()
 	loggedInUser, _ := exec.LoggedInUser()
-	npmrcAudit := detector.NewNPMRCDetector(exec).Detect(ctx, searchDirs, loggedInUser)
+	npmrcAudit := configaudit.NewNPMRCDetector(exec).Detect(ctx, searchDirs, loggedInUser)
 	log.StepDone(time.Since(start))
 
 	// pip config audit — same shape: every pip.conf / pip.ini discovered,
 	// merged effective view, env-var snapshot, and a fixed finding catalog.
 	log.StepStart("Auditing pip configuration")
 	start = time.Now()
-	pipAudit := detector.NewPipConfigDetector(exec).Detect(ctx, loggedInUser)
+	pipAudit := configaudit.NewPipConfigDetector(exec).Detect(ctx, loggedInUser)
 	log.StepDone(time.Since(start))
 
 	// Ensure no nil slices (JSON must emit [] not null)
